+++ x/name/types/params_test.go
@@ -14,14 +14,14 @@ func TestDefaultParams(t *testing.T) {
 	require.NotNil(t, p)
 	require.Equal(t, DefaultMinSegmentLength, p.MinSegmentLength)
 	require.Equal(t, DefaultMaxSegmentLength, p.MaxSegmentLength)
-	require.Equal(t, DefaultMaxSegments, p.MaxNameLevels)
+	require.Equal(t, DefaultMaxNameLevels, p.MaxNameLevels)
 	require.Equal(t, DefaultAllowUnrestrictedNames, p.AllowUnrestrictedNames)
 
-	require.True(t, p.Equal(NewParams(DefaultMaxSegmentLength, DefaultMinSegmentLength, DefaultMaxSegments, DefaultAllowUnrestrictedNames)))
-	require.False(t, p.Equal(NewParams(1, DefaultMinSegmentLength, DefaultMaxSegments, DefaultAllowUnrestrictedNames)))
-	require.False(t, p.Equal(NewParams(DefaultMaxSegmentLength, 1, DefaultMaxSegments, DefaultAllowUnrestrictedNames)))
+	require.True(t, p.Equal(NewParams(DefaultMaxSegmentLength, DefaultMinSegmentLength, DefaultMaxNameLevels, DefaultAllowUnrestrictedNames)))
+	require.False(t, p.Equal(NewParams(1, DefaultMinSegmentLength, DefaultMaxNameLevels, DefaultAllowUnrestrictedNames)))
+	require.False(t, p.Equal(NewParams(DefaultMaxSegmentLength, 1, DefaultMaxNameLevels, DefaultAllowUnrestrictedNames)))
 	require.False(t, p.Equal(NewParams(DefaultMaxSegmentLength, DefaultMinSegmentLength, 1, DefaultAllowUnrestrictedNames)))
-	require.False(t, p.Equal(NewParams(DefaultMaxSegmentLength, DefaultMinSegmentLength, DefaultMaxSegments, false)))
+	require.False(t, p.Equal(NewParams(DefaultMaxSegmentLength, DefaultMinSegmentLength, DefaultMaxNameLevels, false)))
 
 	var p2 *Params
 	require.True(t, p2.Equal(nil))